pyproject.toml
[project]
name = "orgwarden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tech AI", email = "austinweeksmusic@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "repoauditor>=0.1.22",
    "requests>=2.32.3",
    "typer>=0.15.2",
]

[project.scripts]
orgwarden = "orgwarden.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
addopts = "--verbose --capture=no --cov=orgwarden --cov-fail-under=95.0 --cov-report term-missing"
